+++ src/typings/Volunteer.d.ts
@@ -6,3 +6,11 @@ interface Volunteer {
   image: string;
   details: VolunteerDetails | null;
 }
+
+interface RawVolunteer {
+  id: number;
+  user_id: number;
+  name: string;
+  image: string;
+  role: string;
+}
